+++ index.js
@@ -52,6 +52,8 @@ function switchDemo(tab) {
       },
       {
         auto: true,
+        filterVariables: ["--arc-"],
+        defaultOpen: true,
       }
     );
   } else {